requirements.txt
streamlit
requests
google-genai
markdown2